moved layout classes to different namespace
