+++ AnnoDesigner/AnnoDesigner.csproj
@@ -52,10 +52,12 @@
       <DependentUpon>About.xaml</DependentUpon>
     </Compile>
     <Compile Include="Commons.cs" />
+    <Compile Include="model\LayoutVersionContainer.cs" />
     <Compile Include="PresetsLoader\IconLoader.cs" />
     <Compile Include="PresetsLoader\IconMappingPresetsLoader.cs" />
     <Compile Include="model\ICommons.cs" />
     <Compile Include="model\IUpdateHelper.cs" />
+    <Compile Include="model\SavedLayout.cs" />
     <Compile Include="TreeLocalization.cs" />
     <Compile Include="converter\BoolToVisibilityConverter.cs" />
     <Compile Include="Localization.cs" />
